+++ database/migrations/2022_08_16_070844_kategori_final.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('kode_final_kategori')->unique();
             $table->string('kode_utility_kategori');
-            $table->integer('nilai_akhir');
+            $table->decimal('nilai_akhir');
             $table->timestamps();
         });
     }
